+++ src/components/css/cart.css
@@ -68,6 +68,14 @@ input[type="checkbox"]:checked::after {
 
 /* 결제상품 서식 */
 
+.payment_con{
+  height: 290px;
+  background-color: #fff;
+  color: var(--t-color1);
+  margin: var(--b-margin3) 0px;
+  border-radius: var(--b-radius1);
+  padding: 30px 15px;
+}
 .payment_con p{
   margin-bottom: var(--b-margin1);
 }
@@ -75,14 +83,6 @@ input[type="checkbox"]:checked::after {
   float: right;
   font-weight: bold;
   font-size: var(--t-size1);
-}
-.payment_con{
-  height: 290px;
-  background-color: #fff;
-  color: var(--t-color1);
-  margin-top: 30px;
-  border-radius: var(--b-radius1);
-  padding: 30px 15px;
 }
 .p_title{
   font-weight: bold;
@@ -111,6 +111,7 @@ input[type="checkbox"]:checked::after {
 
 
 /* 구매내역 */
+form{width:360px; margin: 0 auto;}
 h2{text-align: center;}
 .payment_list, .payment_com{
   height: 100px;
@@ -129,7 +130,6 @@ h2{text-align: center;}
 }
 
 h3{
-  color: var(--t-color2);
   font-weight: bold;
   padding-bottom: 10px;
   margin-bottom: var(--b-margin2);
