+++ src/components/nav/cart/Purchase.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Payment from './Payment';
-import '../css/cart.css';
+import '../../css/cart.css';
 
 function Purchase() {
   return (
